rocks: unable to install with tarantoolctl rocks on Linux
On any platforms many targets from subdirectories are included into make
command by default. They are not necessary to build driver.{so|dylib}.
EXCLUDE_FROM_ALL option disables all targets at the beginning and only targets
which are explicitly added to mqtt dependencies will be built.

Static build option for libmosquitto set by default in rocks. That means there
is no more need to install libmosquitto manually for successful build from rocks.
But there is still an opportunity to add some other build options in CMake,
if it is needed.

WITH_PIC added as an option. This is related to the CMP0077 policy.
https://cmake.org/cmake/help/latest/policy/CMP0077.html. Without this
change the behaviour is the following. If the module builds statically,
WITH_PIC is firstly set on, but the effect of this will be discarded by
option(WITH_PIC <...> OFF) in third_party/mosquitto/lib/CMakeLists.txt.
This way the build fails.

Fixes #39